Move fastAPI to core_api from embed and update code to support that
